tests added, attendee services modified , location updated

+++ hebtus_crossplatform/lib/models/location.dart
@@ -5,8 +5,8 @@ class Location {
   Location({required this.longitude, required this.latitude});
 
   Location.fromJson(Map<String, dynamic> json)
-      : longitude = json["coordinates"][0],
-        latitude = json["coordinates"][1];
+      : longitude = json["coordinates"][0].toDouble(),
+        latitude = json["coordinates"][1].toDouble();
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = <String, dynamic>{};
